feat: Move `matchRegexOrMinimatch` to string utilities
