+++ lib/util/string.ts
@@ -1,3 +1,6 @@
+import { minimatch } from './minimatch';
+import { regEx } from './regex';
+
 // Return true if the match string is found at index in content
 export function matchAt(
   content: string,
@@ -93,4 +96,24 @@ export function coerceString(
   def?: string,
 ): string {
   return val ?? def ?? '';
+}
+
+export function matchRegexOrMinimatch(input: string, pattern: string): boolean {
+  if (pattern.length > 2 && pattern.startsWith('/') && pattern.endsWith('/')) {
+    try {
+      const regex = regEx(pattern.slice(1, -1));
+      return regex.test(input);
+    } catch (err) {
+      return false;
+    }
+  }
+
+  return minimatch(pattern, { dot: true }).match(input);
+}
+
+export function anyMatchRegexOrMinimatch(
+  input: string,
+  patterns: string[],
+): boolean | null {
+  return patterns.some((pattern) => matchRegexOrMinimatch(input, pattern));
 }
